My take on the gg90 with precalc and some attempts at optimization. This is for DCT with n=32.
Notable is the precalc that occurs at the beginning of the program, before timing starts.

I recall that the precalc tables looked correct; but my program output does not seem to be correct.
I think maybe I got one of the gg90 loops wrong (I combined a couple of for loops).

There are some very rough coments at the end of the source file to indicate what I modified, in steps that I timed each run.

A sample output file is included with the source. The timing values are roughly a third of the original, unoptimized source.